got rid of warning

+++ c/microratchetcrypto/rng.c
@@ -3,7 +3,7 @@
 
 mr_rng_ctx mr_rng_create(mr_ctx mr_ctx)
 {
-	return (mr_rng_ctx)0xffffffff;
+	return (mr_rng_ctx)(size_t)0xffffffff;
 }
 
 mr_result mr_rng_generate(mr_rng_ctx _ctx, uint8_t* output, uint32_t outputsize)
